Install packages with pip only locally
pip will always fetch the most recent version of a package,
regardless of the packages already installed. Ideally, one
would curate a list of pinned versions for each package and
their dependencies, and supply that list to pip whenever a
new package needs to be installed. Instead, we will maintain
two images: Ubuntu 18.04 with CUDA 10 where pip is used, and
Ubuntu 18.04 where pip is not used.