Reduce build time for Intel
Split the package update in two blocks: Ubuntu packages required to
install the Intel compiler are fetched first, then the Intel compiler
is installed, then the remaining Ubuntu packages are installed. This
way, the Intel compiler should remain in the cache, saving 1 to
4 hours of installation time (the amount of time is random), while
still allowing modifications to the list of Ubuntu packages.